Fix usage of size_t vs uint64_t where the other was the intention, add faster convolution functions (the SSE ones don't enforce 16 byte alignment), improve speed of some other ion::Matrix functions, make ion::Matrix::Roi_Fast even faster, add ioN::Matrix::Roi_ReallyFast, improve readability of ion::Convolve and start adding an ion::Convolve2D (not fully tested),

+++ common/ionlib/inc/ionlib/math.h
@@ -35,7 +35,7 @@ namespace ion
 #define FEET_TO_METERS (0.3048)
 
 	double randlf(double low, double high);
-	size_t randull(size_t low, size_t high);
+	uint64_t randull(uint64_t low, uint64_t high);
 	double random_normal_distribution(double low, double high);
 
 	template <class T>
